Moving VinzRunner to scanner instead of scanner/api because it is going to get used everywhere in scanner
